+++ tests/cli/generate/routes/swagger/ErrorsCept.php
@@ -1,0 +1,6 @@
+<?php 
+$I = new CliTester($scenario);
+$I->runShellCommand('bin/rest-deamon generate:routes:swagger dir', false);
+$I->seeInShellOutput('`dir not is directory`');
+
+$I->seeResultCodeIs(1);