+++ tests/cli/generate/routes/swagger/FileCept.php
@@ -1,0 +1,3 @@
+<?php 
+$I = new CliTester($scenario);
+$I->wantTo('perform actions and see result');