+++ tests/_fixtures/routes.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'endpoints' => [
+        '/' => [
+            'name' => 'Root Resource',
+            'handlers' => [
+                'GET' => \Example\Swagger\RootResourceHandler::class,
+            ],
+        ],
+    ],
+];