+++ tests/cli/HelpCept.php
@@ -1,0 +1,5 @@
+<?php 
+$I = new CliTester($scenario);
+$I->runShellCommand('bin/rest-deamon');
+$I->seeInShellOutput('help');
+$I->seeResultCodeIs(0);